@@ -1,4 +1,5 @@
-from typing import IO, Callable
+from collections.abc import Callable
+from typing import IO
 
 from packaging.requirements import Requirement
 
@@ -8,10 +9,10 @@ def sys_info(fid: IO | None = None, developer: bool = False):
     Parameters
     ----------
     fid : file-like | None
-        The file to write to, passed to :func:`print`. Can be None to use
-        :data:`sys.stdout`.
+    The file to write to, passed to :func:`print`. Can be None to use
+    :data:`sys.stdout`.
     developer : bool
-        If True, display information about optional dependencies.
+    If True, display information about optional dependencies.
     """
 
 def _list_dependencies_info(
