@@ -1,4 +1,3 @@
-from _typeshed import Incomplete
 from mne import Info
 from mne.preprocessing import ICA
 from numpy.typing import NDArray
@@ -6,9 +5,10 @@ from numpy.typing import NDArray
 from ..utils._checks import _validate_ica as _validate_ica
 from ..utils._docs import fill_doc as fill_doc
 
+@fill_doc
 def get_topomaps(
     ica: ICA,
-    picks: Incomplete | None = None,
+    picks=None,
     res: int = 64,
     image_interp: str = ...,
     border: float | str = ...,
@@ -19,44 +19,45 @@ def get_topomaps(
     Parameters
     ----------
     ica : ICA
-        MNE `~mne.preprocessing.ICA` decomposition.
+    MNE `~mne.preprocessing.ICA` decomposition.
     picks : int | list of int | slice | None
-        Indices of the independent components (ICs) to select.
-        If an integer, represents the index of the IC to pick.
-        Multiple ICs can be selected using a list of int or a slice.
-        The indices are 0-indexed, so ``picks=1`` will pick the second IC: ``ICA001``.
-        ``None`` (default) will pick all independent components in the order fitted.
+    Indices of the independent components (ICs) to select.
+    If an integer, represents the index of the IC to pick.
+    Multiple ICs can be selected using a list of int or a slice.
+    The indices are 0-indexed, so ``picks=1`` will pick the second IC: ``ICA001``.
+    ``None`` (default) will pick all independent components in the order fitted.
     res : int
-        The resolution of the square topographic map (in pixels).
+    The resolution of the square topographic map (in pixels).
     image_interp : str
-        The image interpolation to be used. All matplotlib options are
-        accepted.
+    The image interpolation to be used. All matplotlib options are
+    accepted.
     border : float | 'mean'
-        Value to extrapolate to on the topomap borders. If ``'mean'`` (default),
-        then each extrapolated point has the average value of its neighbours.
+    Value to extrapolate to on the topomap borders. If ``'mean'`` (default),
+    then each extrapolated point has the average value of its neighbours.
     extrapolate : str
-        Options:
+    Options:
 
-        - ``'box'``
-            Extrapolate to four points placed to form a square encompassing all
-            data points, where each side of the square is three times the range
-            of the data in the respective dimension.
-        - ``'local'`` (default for MEG sensors)
-            Extrapolate only to nearby points (approximately to points closer than
-            median inter-electrode distance). This will also set the
-            mask to be polygonal based on the convex hull of the sensors.
-        - ``'head'`` (default for non-MEG sensors)
-            Extrapolate out to the edges of the clipping circle. This will be on
-            the head circle when the sensors are contained within the head circle,
-            but it can extend beyond the head when sensors are plotted outside
-            the head circle.
+    - ``'box'``
+        Extrapolate to four points placed to form a square encompassing all
+        data points, where each side of the square is three times the range
+        of the data in the respective dimension.
+    - ``'local'`` (default for MEG sensors)
+        Extrapolate only to nearby points (approximately to points closer than
+        median inter-electrode distance). This will also set the
+        mask to be polygonal based on the convex hull of the sensors.
+    - ``'head'`` (default for non-MEG sensors)
+        Extrapolate out to the edges of the clipping circle. This will be on
+        the head circle when the sensors are contained within the head circle,
+        but it can extend beyond the head when sensors are plotted outside
+        the head circle.
 
     Returns
     -------
     topomaps : dict of array of shape (n_components, n_pixels, n_pixels)
-        Dictionary of ICs topographic maps for each channel type.
+    Dictionary of ICs topographic maps for each channel type.
     """
 
+@fill_doc
 def _get_topomap_array(
     data: NDArray[float],
     info: Info,
@@ -70,37 +71,37 @@ def _get_topomap_array(
     Parameters
     ----------
     data : array of shape (n_channels,)
-        The data points used to generate the topographic map.
+    The data points used to generate the topographic map.
     info : Info
-        Instance of `mne.Info` with the montage associated with the ``(n_channels,)``
-        points.
+    Instance of `mne.Info` with the montage associated with the ``(n_channels,)``
+    points.
     res : int
-        The resolution of the square topographic map (in pixels).
+    The resolution of the square topographic map (in pixels).
     image_interp : str
-        The image interpolation to be used. All matplotlib options are
-        accepted.
+    The image interpolation to be used. All matplotlib options are
+    accepted.
     border : float | 'mean'
-        Value to extrapolate to on the topomap borders. If ``'mean'`` (default),
-        then each extrapolated point has the average value of its neighbours.
+    Value to extrapolate to on the topomap borders. If ``'mean'`` (default),
+    then each extrapolated point has the average value of its neighbours.
     extrapolate : str
-        Options:
+    Options:
 
-        - ``'box'``
-            Extrapolate to four points placed to form a square encompassing all
-            data points, where each side of the square is three times the range
-            of the data in the respective dimension.
-        - ``'local'`` (default for MEG sensors)
-            Extrapolate only to nearby points (approximately to points closer than
-            median inter-electrode distance). This will also set the
-            mask to be polygonal based on the convex hull of the sensors.
-        - ``'head'`` (default for non-MEG sensors)
-            Extrapolate out to the edges of the clipping circle. This will be on
-            the head circle when the sensors are contained within the head circle,
-            but it can extend beyond the head when sensors are plotted outside
-            the head circle.
+    - ``'box'``
+        Extrapolate to four points placed to form a square encompassing all
+        data points, where each side of the square is three times the range
+        of the data in the respective dimension.
+    - ``'local'`` (default for MEG sensors)
+        Extrapolate only to nearby points (approximately to points closer than
+        median inter-electrode distance). This will also set the
+        mask to be polygonal based on the convex hull of the sensors.
+    - ``'head'`` (default for non-MEG sensors)
+        Extrapolate out to the edges of the clipping circle. This will be on
+        the head circle when the sensors are contained within the head circle,
+        but it can extend beyond the head when sensors are plotted outside
+        the head circle.
 
     Returns
     -------
     topomap : array of shape (n_pixels, n_pixels)
-        Topographic map array.
+    Topographic map array.
     """
